requirements.txt
web3
requests
yfinance
python-dotenv
